Added lru cache
