--- packages/stitcher/src/presentation/index.ts
@@ -1,5 +1,0 @@
-export * from "./presentation.js";
-
-export * from "./types.js";
-
-export * from "./utils.js";
