--- packages/stitcher/src/presentation/types.ts
@@ -1,6 +1,0 @@
-import type { Variant, MediaPlaylist } from "../parser/index.js";
-
-export type Stream = {
-  variant: Variant;
-  media: MediaPlaylist;
-};
